+++ R/esp_get_nuts.R
@@ -193,9 +193,16 @@ esp_get_nuts <- function(year = "2016",
   }
 
   # Get region id
-
   if (is.null(region)) {
-    nuts_id <- NULL
+    # Get codes from cached data
+    # This overcome the issue on giscoR
+    # https://github.com/rOpenGov/giscoR/issues/57
+    getids <- sf::st_drop_geometry(mapSpain::esp_nuts.sf)
+
+    if (nuts_level != "all"){
+      getids <- getids[getids$LEVL_CODE == nuts_level, ]
+    }
+    nuts_id <- as.character(getids$NUTS_ID)
   } else {
     nuts_id <- esp_hlp_all2nuts(region)